refactor: EditorSection と MonacoEditor を新しいディレクトリに移動
